% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossbasis.R
\name{add_min_dist}
\alias{add_min_dist}
\title{Calculate distance between predicted values and actual data points on a grid}
\usage{
add_min_dist(df, Q_name, L, pred)
}
\arguments{
\item{df}{data frame; model data}

\item{Q_name}{quosure; the name for the Q matrix}

\item{L}{matrix; the actual L matrix}

\item{pred}{data frame; the predicted values}
}
\value{
a tibble
}
\description{
This is inspired by the \code{too.far} argument in \code{plot.gam()}.  It takes
predicted values and adds the distance to the model data.  You can then use
the \code{min_dist} column to filter data for plotting.
}
