% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotate_spei}
\alias{annotate_spei}
\title{Add SPI or SPEI indicators to a ggplot}
\usage{
annotate_spei(p, ...)
}
\arguments{
\item{p}{a ggplot object}

\item{...}{other arguments passed to \code{\link[=annotate]{annotate()}}, e.g. \code{alpha}}
}
\value{
a ggplot object
}
\description{
Add SPI or SPEI indicators to a ggplot
}
