% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eu_dist}
\alias{eu_dist}
\title{Calculate euclidean distance between two points}
\usage{
eu_dist(p1, p2)
}
\arguments{
\item{p1}{numeric vector of length 2; x and y coordinates of point 1}

\item{p2}{numeric vector of length 2; x and y coordinates of point 2}
}
\value{
Euclidean distance between the points
}
\description{
Calculate euclidean distance between two points
}
\examples{
x = c(0,0)
y = c(3,3)
eu_dist(x, y)
}
