% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%^\%}
\alias{\%^\%}
\title{Matrix iteration operator}
\usage{
x \%^\% k
}
\arguments{
\item{x}{a matrix}

\item{k}{a power}
}
\value{
a matrix
}
\description{
A wrapper around \link[matrixcalc:matrix.power]{matrixcalc::matrix.power}. A\%^\%k multiplies matrix A by itself k times.
}
\examples{
A <- matrix(1:9, nrow = 3)
B <- A \%^\% 3
B == A \%*\% A \%*\% A
}
