% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellner_utils.R
\name{mk_K}
\alias{mk_K}
\title{Make Projection Matrix K}
\usage{
mk_K(P_z1z, F_z1z, m = 100, L, U)
}
\arguments{
\item{P_z1z}{kernel function for growth and survival}

\item{m}{number of mesh points}

\item{L}{lower boundary}

\item{U}{upper boundary}

\item{F_ziz}{kernel function for fecundity}
}
\value{
a list with elements K, P, F, and meshpts
}
\description{
Make Projection Matrix K
}
\examples{
\dontrun{
mk_K(m = 100, L = 2.4, U = 200.2)
}
}
