% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearest}
\alias{nearest}
\title{Which element(s) of a vector is (are) nearest to some value?}
\usage{
nearest(x, val)
}
\arguments{
\item{x}{a numeric vector}

\item{val}{a numeric scalar}
}
\value{
a logical vector
}
\description{
For use in dplyr::filter() to get all the rows with the value closest to some target value.
}
\examples{
x <- seq(1, 10, 0.3)
nearest(x, 5)
}
\seealso{
dplyr::near()
}
